+++ src/apps/PhotoViewer.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck: temporary suppression of pre-existing type errors during incremental migration
 import React from 'react';
 import styled from 'styled-components';
 import { useApp } from '../hooks/useApp';